requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-decouple==2.1
django-filter==22.1
django-jet-reboot==1.3.1
django-rest-framework==0.1.0
django-rosetta==0.9.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
feedparser==6.0.10
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.2
kavenegar==1.1.2
Khayyam==3.0.17
Markdown==3.3.7
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
polib==1.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.9
